직관 기록 상세 파일명 수정
