+++ lib/ui/pages/holder/visit_record/detail_page/visit_record_detail_page.dart
@@ -1,6 +1,6 @@
 import 'package:ballkkaye_frontend/_core/style/m_icon.dart';
 import 'package:ballkkaye_frontend/_core/style/m_text.dart';
-import 'package:ballkkaye_frontend/ui/pages/holder/visit_record/detail_page/widgets/visit_record_detail_body.dart';
+import 'package:ballkkaye_frontend/ui/pages/holder/visit_record/detail_page/widgets/detail_body.dart';
 import 'package:flutter/material.dart';
 
 class VisitRecordDetailPage extends StatelessWidget {
@@ -10,7 +10,7 @@ class VisitRecordDetailPage extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: _appbar(context),
-      body: VisitRecordDetailBody(),
+      body: DetailBody(),
       floatingActionButton: FloatingActionButton(
         onPressed: () {
           Navigator.pushNamed(context, "/visit-record/write");
