--- lib/ui/pages/holder/visit_record/detail_page/widgets/visit_record_detail_body.dart
@@ -1,32 +1,0 @@
-import 'package:ballkkaye_frontend/ui/pages/holder/visit_record/detail_page/widgets/visit_record_detail_content.dart';
-import 'package:ballkkaye_frontend/ui/pages/holder/visit_record/detail_page/widgets/visit_record_detail_game.dart';
-import 'package:ballkkaye_frontend/ui/pages/holder/visit_record/detail_page/widgets/visit_record_detail_image.dart';
-import 'package:flutter/material.dart';
-
-class VisitRecordDetailBody extends StatelessWidget {
-  const VisitRecordDetailBody({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return SingleChildScrollView(
-      padding: EdgeInsets.symmetric(horizontal: 16, vertical: 22),
-      child: Card(
-        elevation: 5,
-        child: Padding(
-          padding: EdgeInsets.all(14),
-          child: Column(
-            children: [
-              VisitRecordDetailGame(),
-              SizedBox(height: 19),
-              VisitRecordDetailImage(),
-              SizedBox(height: 19),
-              VisitRecordDetailContent(),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
